split apart embedded jetty standalone from web_client
