+++ odcs-web-client-jetty-standalone/src/main/java/org/opendcs/odcswebclient/jetty/Start.java
@@ -1,4 +1,11 @@
-package start;
+/*
+ * Copyright (c) 2023
+ * United States Army Corps of Engineers - Hydrologic Engineering Center (USACE/HEC)
+ * All Rights Reserved.  USACE PROPRIETARY/CONFIDENTIAL.
+ * Source may not be released without written approval from HEC
+ */
+
+package org.opendcs.odcswebclient.jetty;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
