+++ web_client/build.gradle
@@ -6,14 +6,8 @@ plugins {
 }
 
 dependencies {
-    implementation(libs.jetty.server)
-    implementation(libs.jetty.webapp)
     implementation(libs.servlet.api)
     implementation(libs.websocket)
-}
-
-war {
-    webAppDirName = 'src/main/webapp'
 }
 
 publishing {
